+++ src/components/InstancesList/InstancesList.js
@@ -57,6 +57,7 @@ import {
   isTestEnv,
   handleKeyCommand,
   buildSingleItemQuery,
+  isUserInConsortiumMode,
 } from '../../utils';
 import {
   INSTANCES_ID_REPORT_TIMEOUT,
@@ -727,14 +728,6 @@ class InstancesList extends React.Component {
       label: intl.formatMessage({ id: `ui-inventory.actions.menuSection.sortBy.${option}` }),
     }));
 
-    const sortByOptions = [
-      {
-        value: '',
-        label: intl.formatMessage({ id: 'ui-inventory.actions.menuSection.sortBy.relevance' }),
-      },
-      ...sortOptions,
-    ];
-
     const getSortByValue = () => {
       return this.state.segmentsSortBy.find(x => x.name === segment).sort?.replace('-', '') || '';
     };
@@ -743,18 +736,14 @@ class InstancesList extends React.Component {
       <>
         <MenuSection label={intl.formatMessage({ id: 'ui-inventory.actions' })} id="actions-menu-section">
           <IfPermission perm="ui-inventory.instance.create">
-            <Button
-              buttonStyle="dropdownItem"
-              id="clickable-newinventory"
-              onClick={buildOnClickHandler(this.openCreateInstance)}
-            >
-              <Icon
-                icon="plus-sign"
-                size="medium"
-                iconClassName={css.actionIcon}
-              />
-              <FormattedMessage id="stripes-smart-components.new" />
-            </Button>
+            {this.getActionItem({
+              id: 'clickable-newinventory',
+              icon: checkIfUserInCentralTenant(stripes) ? 'graph' : 'plus-sign',
+              messageId: !isUserInConsortiumMode(stripes)
+                ? 'stripes-smart-components.new'
+                : (checkIfUserInCentralTenant(stripes) ? 'ui-inventory.newSharedRecord' : 'ui-inventory.newLocalRecord'),
+              onClickHandler: buildOnClickHandler(this.openCreateInstance),
+            })}
           </IfPermission>
           {!checkIfUserInCentralTenant(stripes) && (
             <Pluggable
@@ -849,7 +838,7 @@ class InstancesList extends React.Component {
         >
           <Select
             data-testid="sort-by-selection"
-            dataOptions={sortByOptions}
+            dataOptions={sortOptions}
             value={getSortByValue()}
             onChange={setSortedColumn}
           />