+++ src/constants.js
@@ -241,6 +241,7 @@ export const SORT_DIRECTION = {
 export const SORTABLE_SEARCH_RESULT_LIST_COLUMNS = {
   TITLE: 'title',
   CONTRIBUTORS: 'contributors',
+  RELEVANCE: 'relevance',
 };
 
 
@@ -635,4 +636,12 @@ export const RECORD_SOURCE = {
   UC: 'UC',
 };
 
+export const SOURCE_VALUES = {
+  MARC: 'MARC',
+  FOLIO: 'FOLIO',
+  'CONSORTIUM-MARC': 'MARC-shared',
+  'CONSORTIUM-FOLIO': 'FOLIO-shared',
+};
+
 export const CONSORTIUM_PREFIX = 'CONSORTIUM-';
+export const OKAPI_TENANT_HEADER = 'X-Okapi-Tenant';