+++ src/hooks/index.js
@@ -8,3 +8,4 @@ export { default as useInventoryBrowse } from './useInventoryBrowse';
 export { default as useConfigurationQuery } from './useConfigurationQuery';
 export { default as useLastSearchTerms } from './useLastSearchTerms';
 export { default as useLogout } from './useLogout';
+export { default as useUserTenantPermissions } from './useUserTenantPermissions';