+++ src/routes/InstancesRoute.test.js
@@ -149,7 +149,7 @@ describe('InstancesRoute', () => {
     });
 
     it('should have proper list results size', () => {
-      expect(document.querySelectorAll('#pane-results-content .mclRowContainer > [role=row]').length).toEqual(3);
+      expect(document.querySelectorAll('#pane-results-content .mclRowContainer > [role=row]').length).toEqual(4);
     });
 
     it('should render nothing for select row column header', () => {