+++ src/utils.js
@@ -782,6 +782,6 @@ export const isMARCSource = (source) => {
 
 export const isUserInConsortiumMode = stripes => stripes.hasInterface('consortia');
 
-export const checkIfSharedInstance = (stripes, instance) => {
-  return instance.source?.includes(CONSORTIUM_PREFIX) || checkIfUserInCentralTenant(stripes);
+export const checkIfSharedInstance = (stripes, instance = {}) => {
+  return instance?.source?.includes(CONSORTIUM_PREFIX) || checkIfUserInCentralTenant(stripes);
 };